Images of all the pages
this consist of all the page images it consist, the project is made on android studio platform in dart. the project has direct google sign_in option, forgot_password and many more . this app gives result on autism by letting users enter the results they feel like to in the questionaries. the result are calculated and give a number of the autism percentage or number which tells us if the person has autism or not.